refactored rlcard
